+++ adagucserverEC/CConvertGeoJSON.cpp
@@ -93,14 +93,11 @@ void CConvertGeoJSON::drawpolyWithHoles_index(int xMin, int yMin, int xMax, int 
   int IMAGE_RIGHT = xMax;
   int scanLineWidth = IMAGE_RIGHT - IMAGE_LEFT;
 
-  int cntLines = 0;
-  int cntHoleLists = 0;
   // Allocate  scanline
   unsigned short scanline[scanLineWidth];
   // Loop through the rows of the image.
   for (pixelY = IMAGE_TOP; pixelY < IMAGE_BOT; pixelY++) {
 
-    cntLines++;
     for (i = 0; i < scanLineWidth; i++) scanline[i] = CCONVERTGEOJSON_FILL;
     buildNodeList(pixelY, nodes, nodeX, polyCorners, polyXY);
     bubbleSort(nodes, nodeX);
@@ -116,7 +113,6 @@ void CConvertGeoJSON::drawpolyWithHoles_index(int xMin, int yMin, int xMax, int 
 
     for (int h = 0; h < holes; h++) {
       buildNodeList(pixelY, nodes, nodeX, holeCorners[h], holeXY[h]);
-      cntHoleLists++;
       bubbleSort(nodes, nodeX);
       for (i = 0; i < nodes; i += 2) {
         int x1 = nodeX[i] - IMAGE_LEFT;
@@ -323,8 +319,6 @@ void CConvertGeoJSON::addCDFInfo(CDFObject *cdfObject, CServerParams *, BBOX &df
   unsigned short f = CCONVERTGEOJSON_FILL;
   polygonIndexVar->setAttribute("_FillValue", CDF_USHORT, &f, 1);
 
-  //        std::vector<Feature*>::iterator sample=featureMap.begin();
-  //        Feature *sample=featureMap[0];
   int nrFeatures = featureMap.size();
 
   CDF::Dimension *dimFeatures;
@@ -481,7 +475,7 @@ void CConvertGeoJSON::getDimensions(CDFObject *cdfObject, json_value &json, bool
                   json_value fldValue = *fldObject.value;
                   if (fldValue.type == json_string) {
                     CT::string value(fldValue.u.string.ptr);
-                    CDBDebug("[ ] dim[%s]: %s=%s", dimName.c_str(), fldName.c_str(), value.c_str());
+                    // CDBDebug("[ ] dim[%s]: %s=%s", dimName.c_str(), fldName.c_str(), value.c_str());
                     //                      CDBDebug("[%d] prop[%s]=%s", cnt, propName.c_str(), prop.u.string.ptr);
                     //                      CDBDebug("[%d] prop[%s]S =%s", cnt, propName.c_str(),prop.u.string.ptr);
                     if (fldName.equals("units")) {