+++ adagucserverEC/CDataReader.cpp
@@ -825,8 +825,6 @@ int CDataReader::openExtent(CDataSource *dataSource, int mode, int *gridExtent) 
 
 int CDataReader::open(CDataSource *dataSource, int mode, int x, int y, int *gridExtent) {
 
-#define CDATAREADER_DEBUG 1
-  CDBDebug("Open data source");
   // Perform some checks on pointers
   if (dataSource == NULL) {
     CDBError("Invalid dataSource");
@@ -1025,10 +1023,7 @@ int CDataReader::open(CDataSource *dataSource, int mode, int x, int y, int *grid
      * DataPostProc: Here our datapostprocessor comes into action! It needs scale and offset from datasource.
      * This is stage1, only AX+B will be applied to scale and offset parameters
      */
-    CDBDebug("Request date is   %f", requestDate);
     CDataPostProcessor::getCDPPExecutor()->executeProcessors(dataSource, CDATAPOSTPROCESSOR_RUNBEFOREREADING, requestDate);
-  } else {
-    CDBDebug("Skipping POSTPROC!@!!!!!");
   }
 
   if (mode == CNETCDFREADER_MODE_GET_METADATA) {