+++ adagucserverEC/CDataSource.cpp
@@ -357,18 +357,15 @@ int CDataSource::setCFGLayer(CServerParams *_srvParams, CServerConfig::XMLE_Conf
   srvParams = _srvParams;
   cfg = _cfg;
   cfgLayer = _cfgLayer;
-  //    numVariables = cfgLayer->Variable.size();
-  // CDBDebug("Configure layer ");
   datasourceIndex = layerIndex;
+
+  // Make DataObjects for each Variable defined in the Layer.
   for (size_t j = 0; j < cfgLayer->Variable.size(); j++) {
     DataObject *newDataObject = new DataObject();
     newDataObject->variableName.copy(cfgLayer->Variable[j]->value.c_str());
-    if (!cfgLayer->Variable[j]->attr.orgname.empty()) {
-      newDataObject->variableName = cfgLayer->Variable[j]->value.c_str();
-    }
-
     getDataObjectsVector()->push_back(newDataObject);
   }
+
   // Set the layername
   CT::string layerUniqueName;
   if (_layerName == NULL) {
@@ -383,12 +380,12 @@ int CDataSource::setCFGLayer(CServerParams *_srvParams, CServerConfig::XMLE_Conf
     layerName = "";
   layerName.concat(_layerName);
 
+  layerTitle = cfgLayer->Title.size() > 0 && !cfgLayer->Title[0]->value.empty() ? cfgLayer->Title[0]->value.c_str() : layerName.c_str();
+
 #ifdef CDATASOURCE_DEBUG
   CDBDebug("LayerName=\"%s\"", layerName.c_str());
 #endif
   // Defaults to database
-  // At this point, type does not always have a value
-  // CDBError("!!!!LAYERRRRRRRRRRRRRRRRRRRRRRRRRR TYPE IS  %s", cfgLayer->attr.type.c_str());
   dLayerType = CConfigReaderLayerTypeDataBase;
   if (cfgLayer->attr.type.equals("database")) {
     dLayerType = CConfigReaderLayerTypeDataBase;
@@ -470,6 +467,7 @@ CT::string CDataSource::getDimensionValue(int i) { return timeSteps[currentAnima
 int CDataSource::getNumTimeSteps() { return (int)timeSteps.size(); }
 
 const char *CDataSource::getLayerName() { return layerName.c_str(); }
+const char *CDataSource::getLayerTitle() { return layerTitle.c_str(); }
 
 CCDFDims *CDataSource::getCDFDims() {
   if (currentAnimationStep >= int(timeSteps.size())) {
@@ -725,8 +723,7 @@ int CDataSource::makeStyleConfig(CStyleConfiguration *styleConfig, CDataSource *
       s->renderMethod=nearest;
       }
   }*/
-  CT::string styleDump;
-  styleConfig->printStyleConfig(&styleDump);
+  CT::string styleDump = styleConfig->c_str();
   //   #ifdef CDATASOURCE_DEBUG
   //
   //   CDBDebug("styleDump:\n%s",styleDump.c_str());