+++ adagucserverEC/CImageDataWriter.h
@@ -79,6 +79,7 @@ private:
   // int status;
   int animation;
   int nrImagesAdded;
+  CT::string eProfileJson;
 
 public:
 public: