+++ python/python_fastapi_server/routers/covjsonresponse.py
@@ -1,0 +1,6 @@
+from fastapi.responses import JSONResponse
+
+
+class CovJSONResponse(JSONResponse):
+    media_type = "application/prs.coverage+json"
+