+++ tests/functional_test.py
@@ -19,6 +19,7 @@ from AdagucTests.TestMetadataService import TestMetadataService
 from AdagucTests.TestWMSVolScan import TestWMSVolScan
 from AdagucTests.TestWMSPolylineLabel import TestWMSPolylineLabel
 from AdagucTests.TestDataPostProcessor import TestDataPostProcessor
+from AdagucTests.TestWMSTimeHeightProfiles import TestWMSTimeHeightProfiles
 
 suites = []
 TestLoader = unittest.TestLoader
@@ -36,6 +37,7 @@ suites.append(TestLoader().loadTestsFromTestCase(TestMetadataService))
 suites.append(TestLoader().loadTestsFromTestCase(TestWMSVolScan))
 suites.append(TestLoader().loadTestsFromTestCase(TestWMSPolylineLabel))
 suites.append(TestLoader().loadTestsFromTestCase(TestDataPostProcessor))
+suites.append(TestLoader().loadTestsFromTestCase(TestWMSTimeHeightProfiles))
 result = unittest.TextTestRunner(verbosity=2).run(unittest.TestSuite(suites))